Separate pipeline_ingress into a types/client/server file
